查询余额接口开发

+++ pay-external-api/src/main/java/com/tfjt/pay/external/unionpay/api/service/UnionPayApiService.java
@@ -20,5 +20,5 @@ public interface UnionPayApiService {
      * 获取同福母账户当前账户余额
      * @return
      */
-    Result<Long> currentBalance();
+    Result<Integer> currentBalance();
 }
